bridge/requirements.txt
psutil>=5.9.0
pyautogui>=0.9.53
python-xlib>=0.31
Pillow>=9.0.0
numpy>=1.22.0
pyftpdlib>=1.5.6
flask>=2.2.0
websocket-server>=0.6.1
structlog>=22.3.0
prometheus-client>=0.15.0